Add files via upload
This website features high-quality images of the restaurant's food and interior, an intuitive menu, and an online reservation system. The design is optimized for both desktop and mobile devices, ensuring a seamless user experience for website visitors. Through the use of clean and efficient code, the website loads quickly and performs well, providing a smooth browsing experience. Overall, the website design effectively communicates the restaurant's offerings and enhances its online presence, and I am proud to have created it.